# libnetwork: don't use strings.Fields() to improve performance