@thaJeztah 2022-04-21
> although I doubt this string parsing was really a big bottleneck, especially compared to shelling out to iptables 😅 🙈

🙊 🙊 🙊  very "likely" not the big bottleneck no. I stumbled upon this beautiful code, and, well, was curious how much performant the alternative was (bit of a lunchtime playground); the alternative also felt more readable, so I thought it was a win-win 😅 

But yes, definitely lots of things to tackle to make the networking stack perform better.